Added int32 to int64 conversion. Fixed #8.

+++ app/dirtree/dirtree.go
@@ -139,7 +139,7 @@ func getFSBlockSize(path string) (int64, error) {
 	if err := syscall.Statfs(path, &stat); err != nil {
 		return 0, err
 	} else {
-		return stat.Bsize, nil
+		return int64(stat.Bsize), nil
 	}
 }
